refactor: name shared Apple codegen spec
