+++ src/WebView.ios.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useCallback, useImperativeHandle, useRef } from 'rea
 import { Image, View, ImageSourcePropType, HostComponent } from 'react-native';
 import invariant from 'invariant';
 
-import RNCWebView, { Commands, NativeProps } from './RNCWebViewIOSNativeComponent';
+import RNCWebView, { Commands, NativeProps } from './RNCWebViewAppleNativeComponent';
 import RNCWebViewModule from './NativeRNCWebViewModule';
 
 import {
