+++ src/WebView.macos.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef, useCallback, useImperativeHandle, useRef } from 'react';
 import { Image, View, ImageSourcePropType, HostComponent } from 'react-native';
 import invariant from 'invariant';
-import RNCWebView, { Commands, NativeProps } from './RNCWebViewIOSNativeComponent';
+import RNCWebView, { Commands, NativeProps } from './RNCWebViewAppleNativeComponent';
 import RNCWebViewModule from './NativeRNCWebViewModule';
 import {
   defaultOriginWhitelist,
